+++ CHANGELOG.md
@@ -23,10 +23,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `erlang:+/1`
 - Added `lists:append/1` and `lists:append/2`
 - Added `erlang:spawn_monitor/1`, `erlang:spawn_monitor/3`
+- Added `lists:dropwhile/2`.
 
 ### Fixed
 - ESP32: improved sntp sync speed from a cold boot.
 - Utilize reserved `phy_init` partition on ESP32 to store wifi calibration for faster connections.
+- Support for zero count in `lists:duplicate/2`.
 
 ## [0.6.6] - Unreleased